Hided test report if it's not exists.
