+++ odd-platform-ui/src/components/DataEntityDetails/Overview/OverviewContainer.tsx
@@ -6,6 +6,8 @@ import {
   getDataEntityDetailsFetching,
   getDataEntityIsDataset,
 } from 'redux/selectors/dataentity.selectors';
+import { fetchDataSetQualityTestReport } from 'redux/thunks';
+import { getDatasetTestReport } from 'redux/selectors';
 import Overview from './Overview';
 
 interface RouteProps {
@@ -26,8 +28,9 @@ const mapStateToProps = (
   dataEntityId: parseInt(dataEntityId, 10),
   isDataset: getDataEntityIsDataset(state, dataEntityId),
   isDataEntityDetailsFetching: getDataEntityDetailsFetching(state),
+  datasetQualityTestReport: getDatasetTestReport(state, dataEntityId),
 });
 
-const mapDispatchToProps = {};
+const mapDispatchToProps = { fetchDataSetQualityTestReport };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Overview);
