+++ odd-platform-ui/src/components/shared/AppGraph/AppGraphLink/AppGraphLinkStyles.ts
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 export const Path = styled('path')(({ theme }) => ({
   fill: 'none',
   stroke: theme.palette.texts.hint,
-  strokeWidth: 1,
+  strokeWidth: 2,
 }));
 
 export const Arrow = styled('path')(({ theme }) => ({
